+++ src/main/kotlin/com/github/ferdistro/springpostmanrequestgenerator/line/JavaLineMarkerProvider.kt
@@ -1,8 +1,7 @@
-package com.github.ferdistro.springpostmanrequestgenerator
+package com.github.ferdistro.springpostmanrequestgenerator.line
 
 import com.github.ferdistro.springpostmanrequestgenerator.services.PostmanRequestGenerator
 import com.intellij.codeInsight.daemon.LineMarkerInfo
-import com.intellij.codeInsight.daemon.LineMarkerProvider
 import com.intellij.openapi.editor.markup.GutterIconRenderer
 import com.intellij.openapi.util.IconLoader
 import com.intellij.psi.PsiElement
@@ -11,16 +10,16 @@ import com.intellij.psi.PsiTypeElement
 import java.awt.event.MouseEvent
 import javax.swing.Icon
 
-class JsonGeneratorLineMarkerProvider(
+class JavaLineMarkerProvider(
     private val postmanRequestGenerator: PostmanRequestGenerator,
-) : LineMarkerProvider {
+) : PostmanLineMarkerProvider {
 
     //should have no-arg constructor
     @Suppress("UNUSED")
     constructor() : this(PostmanRequestGenerator())
 
     companion object {
-        private val icon: Icon = IconLoader.getIcon("/META-INF/icon.svg", JsonGeneratorLineMarkerProvider::class.java)
+        private val icon: Icon = IconLoader.getIcon("/META-INF/icon.svg", JavaLineMarkerProvider::class.java)
     }
 
     override fun getLineMarkerInfo(element: PsiElement): LineMarkerInfo<*>? {
@@ -46,4 +45,3 @@ class JsonGeneratorLineMarkerProvider(
         return anchor as? PsiTypeElement ?: (method.nameIdentifier ?: method)
     }
 }
-
