+++ src/main/resources/META-INF/plugin-java.xml
@@ -1,0 +1,6 @@
+<idea-plugin>
+    <extensions defaultExtensionNs="com.intellij">
+        <codeInsight.lineMarkerProvider language="JAVA"
+                                        implementationClass="com.github.ferdistro.springpostmanrequestgenerator.line.JavaLineMarkerProvider"/>
+    </extensions>
+</idea-plugin>
