+++ src/main/resources/META-INF/plugin-kotlin.xml
@@ -1,0 +1,6 @@
+<idea-plugin>
+    <extensions defaultExtensionNs="com.intellij">
+        <codeInsight.lineMarkerProvider language="KOTLIN"
+                                        implementationClass="com.github.ferdistro.springpostmanrequestgenerator.line.KotlinLineMarkerProvider"/>
+    </extensions>
+</idea-plugin>
