+++ src/main/resources/META-INF/plugin.xml
@@ -1,18 +1,21 @@
 <idea-plugin>
     <id>com.github.ferdistro.springpostmanrequestgenerator</id>
-    <name>spring-postman-request-generator</name>
-    <vendor>FerdiStro</vendor>
+    <name>Generate Spring Postman requests</name>
+    <vendor>alex_let</vendor>
 
     <depends>com.intellij.modules.platform</depends>
     <depends>com.intellij.modules.java</depends>
 
-    <extensions defaultExtensionNs="com.intellij">
-        <codeInsight.lineMarkerProvider language="JAVA" implementationClass="com.github.ferdistro.springpostmanrequestgenerator.JsonGeneratorLineMarkerProvider"/>
-    </extensions>
-    <extensions defaultExtensionNs="com.intellij">
-        <codeInsight.lineMarkerProvider language="KOTLIN" implementationClass="com.github.ferdistro.springpostmanrequestgenerator.JsonGeneratorLineMarkerProvider"/>
-    </extensions>
+    <depends optional="true" config-file="plugin-java.xml">com.intellij.java</depends>
+    <depends optional="true" config-file="plugin-kotlin.xml">org.jetbrains.kotlin</depends>
+
+    <extensionPoints>
+        <extensionPoint name="postmanLineMarkerProvider"
+                        interface="com.github.ferdistro.springpostmanrequestgenerator.line.PostmanLineMarkerProvider"
+                        dynamic="true"/>
+    </extensionPoints>
+
     <extensions defaultExtensionNs="org.jetbrains.kotlin">
-        <supportsKotlinPluginMode supportsK1="true" supportsK2="true" />
+        <supportsKotlinPluginMode supportsK1="true" supportsK2="true"/>
     </extensions>
 </idea-plugin>
